gateway/requirements.txt
# 统一接入网关依赖

# Web 服务
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTTP 客户端
httpx==0.27.0

# 认证
pyjwt==2.9.0

# Redis (限流)
redis==5.2.0

# 配置管理
pyyaml==6.0.2
pydantic==2.9.2

# 测试
pytest==8.3.3
pytest-asyncio==0.24.0
